Making app DPI aware: - All previous adjustments were made from old pixels at 96 dpi to new pixels at any arbitrary dpi. This is not very elegant, it should better be conversion from mm to actual pixels, at current dpi. - GetWindowDpmm: new function to return the dot per millimeters, and not in weird imperial units! - MmToPixels and PixelsToMm: handy conversions functions. - Removed all but one RescaleSizeDpi, which could still be useful; removed all RescaleFontDpi. - Manually calculated the old pixels @ 96dpi to corresponding mm, replaced all constants with new values, and calling MmToPixels:   - TSuperGauge window size.   - All window sizes. Note that new values in mm have been rounded to closest "nice looking" numbers.   - Mouse movements constants. Note that the old values in pixels have been reinstated, in comments, just in case we need them back.   - Cursor Hints size.   - Color maps size.   - Fonts sizes.   - Button separators. - Other parts that were updated:   - Line width in tracks and frequency displays.   - Color Table drawing in Frequency Display, as it does not use the ColorTable.Draw method(?)

+++ Src/Views/TInverseView.cpp
@@ -2831,8 +2831,8 @@ if ( RButtonDown ) {
 
         if ( MouseAxis == MouseAxisHorizontal ) {
 
-            if ( dx > 0 )   ScalingLevel /= 1 + (double) min ( adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
-            else            ScalingLevel *= 1 + (double) min ( adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
+            if ( dx > 0 )   ScalingLevel /= 1 + (double) min ( (double) adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
+            else            ScalingLevel *= 1 + (double) min ( (double) adx, MouseMoveScale ) / MouseMoveScale * ( adx > MouseMoveScaleFast ? 1.0 : 0.3 );
 
             SetScaling ( ScalingLevel );
             }
